# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Shuttle Solution API for 42 Irbid Hackathon
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: Get available trips for tomorrow
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a single trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/trips/{id}/confirm:
    post:
      operationId: confirmTrip
      tags: [trips]
      summary: Manually confirm a trip (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /admin/trips/{id}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a trip (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        departureTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, canceled]
        totalSeats:
          type: integer
        bookedSeats:
          type: integer
        availableSeats:
          type: integer
        minBookingsToConfirm:
          type: integer
        direction:
          type: string
          enum: [to_school, from_school]
        createdAt:
          type: string
      required:
        - id
        - date
        - departureTime
        - status
        - totalSeats
        - bookedSeats
        - availableSeats
        - minBookingsToConfirm
        - direction
        - createdAt
...